Add files via upload
Description: Recursion Program. Accepts array of size 5 or above from the keyboard. Using this size, the array is then populated with random integer values between 10 and 375. The array is displayed, then displayed in reverse. Next creating a function using recursion that displays squares of integers in ascending order, starting from 1 to the first number in the array. Then creating another function that uses recursion to raise a number to a power. This function takes two arguments, the number to be raised to the power is the last number in the array and the power is 2. Lastly, creating a function that uses recursion in order to determine whether or not every number in the array is a prime number.